[lambda] move lambda functions into subfolders under stream_alert
