+++ rules/sample_matchers.py
@@ -13,7 +13,7 @@ You can also supply multiple matchers for many common scenarios:
           matchers=['prod', 'itx_corp'], sink=['csirt-pagerduty'])
 """
 from helpers.base import in_set, last_hour
-from stream_alert.rules_engine import StreamRules
+from stream_alert.rule_processor.rules_engine import StreamRules
 
 matcher = StreamRules.matcher
