+++ rules/sample_rules.py
@@ -1,5 +1,5 @@
 from helpers.base import in_set, last_hour
-from stream_alert.rules_engine import StreamRules
+from stream_alert.rule_processor.rules_engine import StreamRules
 
 rule = StreamRules.rule
