--- stream_alert/__init__.py
@@ -1,1 +1,0 @@
-__version__ = '1.2.0'
