+++ stream_alert/alert_processor/__init__.py
