+++ stream_alert/alert_processor/encrypted_credentials/slack
