+++ stream_alert/alert_processor/main.py
