+++ stream_alert/rule_processor/__init__.py
@@ -1,0 +1,1 @@
+__version__ = '1.2.0'
