+++ stream_alert/rule_processor/config.py
