+++ stream_alert/rule_processor/parsers.py
