+++ stream_alert/rule_processor/pre_parsers.py
