+++ stream_alert/rule_processor/sink.py
