+++ test/unit/test_rules_engine.py
@@ -19,10 +19,10 @@ import json
 
 from nose.tools import assert_equal, assert_not_equal, nottest
 
-from stream_alert.classifier import StreamPayload, StreamClassifier
-from stream_alert.pre_parsers import StreamPreParsers
-from stream_alert.config import load_config
-from stream_alert.rules_engine import StreamRules
+from stream_alert.rule_processor.classifier import StreamPayload, StreamClassifier
+from stream_alert.rule_processor.pre_parsers import StreamPreParsers
+from stream_alert.rule_processor.config import load_config
+from stream_alert.rule_processor.rules_engine import StreamRules
 
 rule = StreamRules.rule
 matcher = StreamRules.matcher
